cancel timers before defuncting connection
This allows us to revert some of the fix to PYTHON-367 without the timer
heap growing out of control as described on that ticket. We cancel
before defuncting to ensure that the timer doesn't hang around while
running callbacks. Between this change and the delegation of callbacks
to a secondary thread (see commit 51923029, part of the original
PYTHON-367 fix), this commit should not cause a regression on
PYTHON-367.

Note that we don't cancel timers before defuncting and retrying -- we're
not done with the logical request and still want that timer to fire.

Now that timers are started on ResponseFuture initialization, we
actually use the value of timeout in the tests, so we no longer pass
that argument, allowing the test to use the default value.